feat: support --max-warnings in the cli
