+++ .changeset/rotten-camels-deny.md
@@ -1,0 +1,5 @@
+---
+'@backstage/cli': patch
+---
+
+Support `--max-warnings` flag for package linting
